Move Binary APIs from Buffers.dll to Memory.dll.
