+++ src/System.Buffers/System.Buffers.sln
@@ -20,8 +20,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{E107E9C1-E89
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ref", "ref", "{2E666815-2EDB-464B-9DF6-380BF4789AD4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "System.Buffers.Performance.Tests", "tests\Performance\System.Buffers.Performance.Tests.csproj", "{8C1FA65F-4AE7-4627-8C2B-DE5EDF779DB5}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,10 +38,6 @@ Global
 		{11AE73F7-3532-47B9-8FF6-B4F22D76456C}.Debug|Any CPU.Build.0 = netstandard-Debug|Any CPU
 		{11AE73F7-3532-47B9-8FF6-B4F22D76456C}.Release|Any CPU.ActiveCfg = netstandard-Release|Any CPU
 		{11AE73F7-3532-47B9-8FF6-B4F22D76456C}.Release|Any CPU.Build.0 = netstandard-Release|Any CPU
-		{8C1FA65F-4AE7-4627-8C2B-DE5EDF779DB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8C1FA65F-4AE7-4627-8C2B-DE5EDF779DB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8C1FA65F-4AE7-4627-8C2B-DE5EDF779DB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8C1FA65F-4AE7-4627-8C2B-DE5EDF779DB5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -52,9 +46,5 @@ Global
 		{62E2AD5F-C8D0-45FB-B6A5-AED2C77F198C} = {1A2F9F4A-A032-433E-B914-ADD5992BB178}
 		{2ADDB484-6F57-4D71-A3FE-A57EC6329A2B} = {E107E9C1-E893-4E87-987E-04EF0DCEAEFD}
 		{11AE73F7-3532-47B9-8FF6-B4F22D76456C} = {2E666815-2EDB-464B-9DF6-380BF4789AD4}
-		{8C1FA65F-4AE7-4627-8C2B-DE5EDF779DB5} = {1A2F9F4A-A032-433E-B914-ADD5992BB178}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B1D0CB97-AED2-48E3-A49C-7C8CD75D3FE6}
 	EndGlobalSection
 EndGlobal
