+++ src/System.Memory/src/System/Buffers/Binary/ReaderBigEndian.cs
