+++ src/System.Memory/src/System/Buffers/Binary/Writer.cs
@@ -16,10 +16,17 @@ namespace System.Buffers.Binary
         public static void WriteMachineEndian<T>(ref Span<byte> buffer, T value)
             where T : struct
         {
-            if (BinaryHelpers.IsReferenceOrContainsReferences<T>())
+#if netcoreapp
+            if (RuntimeHelpers.IsReferenceOrContainsReferences<T>())
             {
-                Environment.FailFast($"Cannot write the non-blittable type <{typeof(T).Name}> into the span.");
+                throw new ArgumentException(SR.Format(SR.Argument_InvalidTypeWithPointersNotSupported, typeof(T)));
             }
+#else
+            if (SpanHelpers.IsReferenceOrContainsReferences<T>())
+            {
+                ThrowHelper.ThrowArgumentException_InvalidTypeWithPointersNotSupported(typeof(T));
+            }
+#endif
             if ((uint)Unsafe.SizeOf<T>() > (uint)buffer.Length)
             {
                 throw new ArgumentOutOfRangeException();
@@ -35,10 +42,17 @@ namespace System.Buffers.Binary
         public static bool TryWriteMachineEndian<T>(ref Span<byte> buffer, T value)
             where T : struct
         {
-            if (BinaryHelpers.IsReferenceOrContainsReferences<T>())
+#if netcoreapp
+            if (RuntimeHelpers.IsReferenceOrContainsReferences<T>())
             {
-                Environment.FailFast($"Cannot write the non-blittable type <{typeof(T).Name}> into the span.");
+                throw new ArgumentException(SR.Format(SR.Argument_InvalidTypeWithPointersNotSupported, typeof(T)));
             }
+#else
+            if (SpanHelpers.IsReferenceOrContainsReferences<T>())
+            {
+                ThrowHelper.ThrowArgumentException_InvalidTypeWithPointersNotSupported(typeof(T));
+            }
+#endif
             if (Unsafe.SizeOf<T>() > (uint)buffer.Length)
             {
                 return false;
