+++ src/System.Memory/tests/Binary/BinaryReaderUnitTests.cs
@@ -7,6 +7,7 @@ using System.Runtime.InteropServices;
 using Xunit;
 
 using static System.Buffers.Binary.BinaryPrimitives;
+using static System.TestHelpers;
 
 namespace System.Buffers.Binary.Tests
 {
