+++ src/System.Memory/tests/Performance/Perf.Span.BinaryReadAndWrite.cs
@@ -7,6 +7,7 @@ using Xunit;
 using System.Net;
 
 using static System.Buffers.Binary.BinaryPrimitives;
+using static System.TestHelpers;
 
 namespace System.Buffers.Binary.Tests
 {
